requirements.txt
flask>=3.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
